create standalone auth module
